miApiJWT/requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
bcrypt<4.0.0
python-jose[cryptography]
pydantic
python-multipart